// src/app/(home)/__tests__/types.ts
export enum MockedChannelInfo {
    searchChannel1 = 'search-channel-id-1',
    searchChannel2 = 'search-channel-id-2',
    categoryChannel1 = 'category-channel-id-1',
    categoryChannel2 = 'category-channel-id-2',
}

export enum MockedVideoInfo {
    searchVideo1 = 'search-video-id-1',
    searchVideo2 = 'search-video-id-2',
    categoryVideo1 = 'category-video-id-1',
    categoryVideo2 = 'category-video-id-2',
}